requirements.txt
# Browser automation
selenium>=4.27.1
webdriver-manager>=4.0.0

# HTTP requests
requests>=2.32.4

# Retry logic
tenacity>=8.2.0

# Data validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# CLI framework
click>=8.1.0
rich>=13.0.0

# Utilities
python-dateutil>=2.8.0
